Added inital support for "www.royalroad.com"
Though there is now support, it could be better. Chapter URLs are
scrapped from the contents page, and stored in $story->metadata. For
some reason, the entire $story structure is echoed to the client with
each heartbeat.

So there's probably a better way to do this that wouldn't result in so
much network traffic. But this works, so it's a starting point.